restructure and added profiler
